+++ fairseq/criterions/ctc.py
@@ -12,7 +12,9 @@ from typing import Optional
 
 import torch
 import torch.nn.functional as F
-from fairseq import metrics, utils
+
+from fairseq import utils
+from fairseq.logging import metrics
 from fairseq.criterions import FairseqCriterion, register_criterion
 from fairseq.dataclass import FairseqDataclass
 from fairseq.data.data_utils import post_process
