+++ fairseq/criterions/label_smoothed_cross_entropy.py
@@ -7,7 +7,8 @@ import math
 from dataclasses import dataclass, field
 
 import torch
-from fairseq import metrics, utils
+from fairseq import utils
+from fairseq.logging import metrics
 from fairseq.criterions import FairseqCriterion, register_criterion
 from fairseq.dataclass import FairseqDataclass
 from omegaconf import II
