+++ fairseq/criterions/label_smoothed_cross_entropy_with_rdrop.py
@@ -8,7 +8,8 @@ from dataclasses import dataclass, field
 
 import torch
 
-from fairseq import metrics, utils
+from fairseq import utils
+from fairseq.logging import metrics
 from fairseq.criterions import register_criterion
 from fairseq.criterions.label_smoothed_cross_entropy import (
     LabelSmoothedCrossEntropyCriterion,
