+++ fairseq/criterions/model_criterion.py
@@ -9,7 +9,8 @@ from typing import Dict, List
 
 import torch
 
-from fairseq import metrics, utils
+from fairseq import utils
+from fairseq.logging import metrics
 from fairseq.criterions import FairseqCriterion, register_criterion
 from fairseq.dataclass import FairseqDataclass
 from fairseq.logging.meters import safe_round
