+++ fairseq/criterions/sentence_ranking.py
@@ -7,7 +7,8 @@ import math
 
 import torch
 import torch.nn.functional as F
-from fairseq import metrics, utils
+from fairseq import utils
+from fairseq.logging import metrics
 from fairseq.criterions import FairseqCriterion, register_criterion
