+++ fairseq/tasks/multilingual_translation.py
@@ -10,7 +10,8 @@ from collections import OrderedDict
 from argparse import ArgumentError
 
 import torch
-from fairseq import metrics, options, utils
+from fairseq import options, utils
+from fairseq.logging import metrics
 from fairseq.data import (
     Dictionary,
     LanguagePairDataset,
