+++ fairseq/tasks/translation.py
@@ -13,7 +13,8 @@ from argparse import Namespace
 from omegaconf import II
 
 import numpy as np
-from fairseq import metrics, utils
+from fairseq import utils
+from fairseq.logging import metrics
 from fairseq.data import (
     AppendTokenDataset,
     ConcatDataset,
